requirements.txt
tensorflow
opencv-python
pillow
numpy
gradio
keras
h5py
